+++ README.txt.txt
@@ -1,1 +1,3 @@
-Foo
+Fix for Koalefants' Family mod for Don't Starve.
+
+Readded Missing File that caused Crashes.
