+++ Working Old Version/workshop-354232464 old/images/modicon.xml
@@ -1,0 +1,3 @@
+<Atlas><Texture filename="modicon.tex" />
+    <Elements><Element name="modicon.tex" u1="0" u2="1" v1="0" v2="1" />
+</Elements></Atlas>
