+++ Working Old Version/workshop-354232464 old/modinfo.lua
@@ -1,0 +1,10 @@
+name = "Koalefants Family"
+description = "This modification adds baby koalefants. They can grow up in the herd. Enjoy!"
+author = "Mr.Hastings, Fidooop,\n and Brainy"
+version = "1.6"
+dont_starve_compatible = true
+reign_of_giants_compatible = true
+forumthread = ""
+api_version = 6
+icon_atlas = "images/modicon.xml"
+icon = "modicon.tex"
