+++ Working Old Version/workshop-354232464 old/scripts/brains/baby_koalefantbrain.lua
@@ -1,0 +1,56 @@
+require "behaviours/wander"
+require "behaviours/faceentity"
+require "behaviours/panic"
+require "behaviours/runaway"
+require "behaviours/follow"
+
+local MIN_FOLLOW_DIST = 2
+local TARGET_FOLLOW_DIST = 3
+local MAX_FOLLOW_DIST = 8
+
+local MAX_CHASE_TIME = 6
+local WANDER_DIST = 20
+
+
+local RUN_AWAY_DIST = 6
+local STOP_RUN_AWAY_DIST = 12
+local START_FACE_DIST = 10
+local KEEP_FACE_DIST = 15
+
+local function GetFaceTargetFn(inst)
+    local target = GetClosestInstWithTag("player", inst, START_FACE_DIST)
+    if target and not target:HasTag("notarget") then
+        return target
+    end
+end
+
+local function KeepFaceTargetFn(inst, target)
+    return inst:GetDistanceSqToInst(target) <= KEEP_FACE_DIST * KEEP_FACE_DIST and not target:HasTag("notarget")
+end
+
+local function ShouldRunAway(guy)
+	return guy:HasTag("character") and not guy:HasTag("notarget")
+end
+
+local KoalefantBrain = Class(Brain, function(self, inst)
+    Brain._ctor(self, inst)
+end)
+
+function KoalefantBrain:OnStart()
+    
+    local root = PriorityNode(
+    {
+        WhileNode( function() return self.inst.components.health.takingfiredamage end, "OnFire", Panic(self.inst)),
+        SequenceNode{
+            FaceEntity(self.inst, GetFaceTargetFn, KeepFaceTargetFn, 0.5),
+            RunAway(self.inst, ShouldRunAway, RUN_AWAY_DIST, STOP_RUN_AWAY_DIST)
+        },
+        FaceEntity(self.inst, GetFaceTargetFn, KeepFaceTargetFn),
+		Follow(self.inst, function() return self.inst.components.follower and self.inst.components.follower.leader end, MIN_FOLLOW_DIST, TARGET_FOLLOW_DIST, MAX_FOLLOW_DIST),
+        Wander(self.inst)
+    }, .25)
+    
+    self.bt = BT(self.inst, root)
+end
+
+return KoalefantBrain
